Move UserSerializer to serializers module
Refs LINK-1122
